Complete animations whether or not cascade is done
Fixes #12554
